requirements.txt
PyMySQL==1.1.1
python-dotenv==1.1.0
SQLAlchemy==2.0.39
tk==0.1.0
typing_extensions==4.12.2
